@@ -2,7 +2,7 @@ name: scholarx
 description: A new Flutter project.
 version: 1.0.0+1
 environment: 
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.17.0 <3.0.0'
 dependencies: 
   flutter: 
     sdk: flutter
