@@ -2,12 +2,16 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:firebase_core/firebase_core.dart';
+import 'package:scholarx/firebase_options.dart';
+import 'package:scholarx/presentation/sign_up_page/auth_page.dart';
 import 'core/app_export.dart';
 
 var globalMessengerKey = GlobalKey<ScaffoldMessengerState>();
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  await Firebase.initializeApp();
+  await Firebase.initializeApp(
+    options: DefaultFirebaseOptions.currentPlatform,
+  );
   SystemChrome.setPreferredOrientations([
     DeviceOrientation.portraitUp,
   ]);
@@ -18,6 +22,7 @@ void main() async {
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({super.key});
   @override
   Widget build(BuildContext context) {
     
@@ -27,7 +32,7 @@ class MyApp extends StatelessWidget {
           theme: theme,
           title: 'scholarx',
           debugShowCheckedModeBanner: false,
-          initialRoute: AppRoutes.jumpstartPageScreen,
+          initialRoute: AppRoutes.authPage,
           routes: AppRoutes.routes,
         );
       },
